Remove separate calculations of label and time column sizes and move sizing function to tableGraph.js
